AUSM working, BC implemented, but small issue in Nk flux

+++ BC.h
@@ -208,27 +208,27 @@ void BC(
 		{
 			if(aline.find("InletDensity")!=string::npos)
 			{
-				InletDensity = stoi (aline.substr(aline.find("=")+1));
+				InletDensity = stod (aline.substr(aline.find("=")+1));
 			}
 			else if(aline.find("InletXVelocity")!=string::npos)
 			{
-				InletXVelocity = stoi (aline.substr(aline.find("=")+1));
+				InletXVelocity = stod (aline.substr(aline.find("=")+1));
 			}
 			else if(aline.find("InletYVelocity")!=string::npos)
 			{
-				InletYVelocity = stoi (aline.substr(aline.find("=")+1));
+				InletYVelocity = stod (aline.substr(aline.find("=")+1));
 			}
 			else if(aline.find("InletZVelocity")!=string::npos)
 			{
-				InletZVelocity = stoi (aline.substr(aline.find("=")+1));
+				InletZVelocity = stod (aline.substr(aline.find("=")+1));
 			}
 			else if(aline.find("InletStaticPressure")!=string::npos)
 			{
-				InletStaticPressure = stoi (aline.substr(aline.find("=")+1));
+				InletStaticPressure = stod (aline.substr(aline.find("=")+1));
 			}
 			else if(aline.find("ExitStaticPressure")!=string::npos)
 			{
-				ExitStaticPressure = stoi (aline.substr(aline.find("=")+1));
+				ExitStaticPressure = stod (aline.substr(aline.find("=")+1));
 			}
 			else if (aline.find("BoundaryConditionati0")!=string::npos)
 			{
